pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []   # no Python packages to install — this is an app, not a library

[project]
name = "finagent"
version = "1.0.0"
description = "Autonomous financial research system — LangGraph + FastAPI + Gemini"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langgraph>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-core>=0.2.0",
    "google-generativeai>=0.7.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "qdrant-client>=1.9.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "pydantic>=2.7.0",
    "edgartools>=5.23",
    "yfinance>=0.2.36",
    "redis>=5.0.0",
    "upstash-redis>=1.0.0",
    "langchain-mcp-adapters>=0.1.8",
    "mcp>=1.0.0",
    "langfuse>=2.0.0",
    "pandas-ta>=0.3.14b ; python_version >= '3.12'",
    "quantstats>=0.0.62",
    "pyportfolioopt>=1.5.5",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

# ---------------------------------------------------------------------------
# Dev dependencies (uv current syntax — replaces deprecated [tool.uv])
# ---------------------------------------------------------------------------
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

# ---------------------------------------------------------------------------
# uv — treat this as a virtual project (app, not an installable library)
# No [build-system] needed; uv won't try to build/install the package itself.
# ---------------------------------------------------------------------------
[tool.uv]
package = false

[tool.uv.sources]
# CPU-only torch — keeps the image ~4 GB smaller than the CUDA build.
# Delete this block if you have a GPU and want CUDA.
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --show-capture=log"
log_cli = true
log_cli_level = "INFO"
log_format = "%(asctime)s %(levelname)s %(name)s: %(message)s"

# ---------------------------------------------------------------------------
# ruff (optional linter)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
